# docs/api/openapi.yaml
components:
  schemas:
    AccessRefreshBundle:
      additionalProperties: false
      description: Native-client token bundle returned by login + refresh. Web cookie-only callers see only `user`.
      properties:
        refreshToken:
          description: Refresh token (`hlr_<64hex>`); only present for native-policy callers.
          type: string
        refreshTokenExpiresAt:
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
          type: string
        token:
          description: Access token (`hlk_<64hex>`).
          type: string
        tokenExpiresAt:
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
          type: string
        user:
          additionalProperties: false
          properties:
            id:
              type: string
            username:
              type: string
          required:
            - id
            - username
          type: object
      required:
        - token
        - tokenExpiresAt
      type: object
...
    ErrorEnvelope:
      additionalProperties: false
      description: "Standard error response: data is null, error is human prose."
      properties:
        data:
          type: "null"
        error:
          type: string
        meta:
          additionalProperties: false
          properties:
            errorCode:
              type: string
            requestId:
              type: string
          type: object
      required:
        - data
        - error
      type: object
...
    LoginResponse:
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/AccessRefreshBundle"
        error:
          type: "null"
        meta:
          additionalProperties: false
          properties:
            requestId:
              type: string
          type: object
      required:
        - data
        - error
      type: object
...
    PasskeyLoginVerifyRequest:
      description: Passkey assertion verification. Same native-client token issuance as the password path.
      properties:
        challengeId:
          description: Server-issued WebAuthn challenge id.
          minLength: 1
          type: string
        credential:
          description: SimpleWebAuthn-style assertion response payload.
          properties:
            authenticatorAttachment:
              enum:
                - platform
                - cross-platform
              type: string
            clientExtensionResults:
              additionalProperties: {}
              propertyNames:
                type: string
              type: object
            id:
              type: string
            rawId:
              type: string
            response:
              additionalProperties: {}
              propertyNames:
                type: string
              type: object
            type:
              const: public-key
              type: string
          required:
            - id
            - rawId
            - type
            - response
          type: object
      required:
        - challengeId
        - credential
      type: object
    PasskeyLoginVerifyResponse:
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/AccessRefreshBundle"
        error:
          type: "null"
        meta:
          additionalProperties: false
          properties:
            requestId:
              type: string
          type: object
      required:
        - data
        - error
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: hlk_<64hex>
      description: Native-client API token. Use the `token` field returned by /api/auth/login on a native client.
      scheme: bearer
      type: http
    cookieAuth:
      description: Browser session cookie — set by /api/auth/login.
      in: cookie
      name: healthlog_session
      type: apiKey
info:
  contact:
    name: HealthLog
    url: https://healthlog.dev
  description: Self-hosted personal-health-tracking PWA — public API surface for the iOS native client and external ingest.
  license:
    name: AGPL-3.0
    url: https://github.com/MBombeck/HealthLog/blob/main/LICENSE
  title: HealthLog API
  version: 1.4.23
openapi: 3.1.0
paths:
  /api/auth/login:
    post:
      description:
        "Browser callers receive a session cookie. Native callers (X-Client-Type: native or HealthLog-iOS UA
        prefix) additionally receive a paired access + refresh token."
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  minLength: 1
                  type: string
                password:
                  minLength: 1
                  type: string
              required:
                - email
                - password
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          description: Login succeeded.
        "401": &a1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Authentication required or invalid credentials.
        "422": &a2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Request validation failed.
        "429": &a3
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Rate limit exceeded.
      summary: Email-or-username login (password)
      tags:
        - Auth
...
  /api/auth/passkey/login-verify:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyLoginVerifyRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyLoginVerifyResponse"
          description: Assertion verified — session + optional bearer issued.
        "401": *a1
        "422": *a2
        "429": *a3
      summary: Passkey assertion verification
      tags:
        - Auth
...
servers:
  - description: Production
    url: https://healthlog.bombeck.io
  - description: Local dev
    url: http://localhost:3000
...